# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bug Bounty Pro API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scan-jobs:
    get:
      operationId: listScanJobs
      tags: [scans]
      summary: List all scan jobs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, running, completed, failed, cancelled]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of scan jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanJobListResponse"

    post:
      operationId: createScanJob
      tags: [scans]
      summary: Create a new scan job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScanJobRequest"
      responses:
        "201":
          description: Scan job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanJob"
        "400":
          description: Bad request
        "429":
          description: Rate limit exceeded
  /scan-jobs/{id}:
    get:
      operationId: getScanJob
      tags: [scans]
      summary: Get scan job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scan job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanJob"
        "404":
          description: Not found

    delete:
      operationId: deleteScanJob
      tags: [scans]
      summary: Delete a scan job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /scan-jobs/{id}/findings:
    get:
      operationId: getScanJobFindings
      tags: [scans]
      summary: Get findings for a scan job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low, info]
        - name: validation_status
          in: query
          schema:
            type: string
            enum: [real, false_positive, informational, pending]
      responses:
        "200":
          description: List of findings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Finding"
  /scan-jobs/{id}/attack-surface:
    get:
      operationId: getScanJobAttackSurface
      tags: [scans]
      summary: Get attack surface graph for a scan job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Attack surface graph data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttackSurfaceGraph"
...
  /targets:
    get:
      operationId: listTargets
      tags: [targets]
      summary: List all scanned targets
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of targets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TargetListResponse"
  /targets/{id}:
    get:
      operationId: getTarget
      tags: [targets]
      summary: Get target by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Target details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Target"
        "404":
          description: Not found
...
components:
  schemas:
...
    ScanJob:
      type: object
      properties:
        id:
          type: string
        target_url:
          type: string
        scan_profile:
          type: string
          enum: [quick, standard, deep]
        status:
          type: string
          enum: [queued, running, completed, failed, cancelled]
        progress:
          type: integer
        created_at:
          type: string
        started_at:
          type: string
        completed_at:
          type: string
        findings_count:
          type: integer
        critical_count:
          type: integer
        high_count:
          type: integer
        medium_count:
          type: integer
        low_count:
          type: integer
        info_count:
          type: integer
        risk_score:
          type: number
        ai_summary:
          type: string
        validation_enabled:
          type: boolean
        fuzzing_enabled:
          type: boolean
        bug_bounty_mode:
          type: boolean
        authorization_acknowledged:
          type: boolean
        scanner_engines:
          type: array
          items:
            type: string
        error_message:
          type: string
      required:
        - id
        - target_url
        - scan_profile
        - status
        - progress
        - created_at
        - findings_count
        - critical_count
        - high_count
        - medium_count
        - low_count
        - info_count
        - risk_score
        - validation_enabled
        - fuzzing_enabled
        - bug_bounty_mode
        - authorization_acknowledged
    CreateScanJobRequest:
      type: object
      properties:
        target_url:
          type: string
        scan_profile:
          type: string
          enum: [quick, standard, deep]
        validation_enabled:
          type: boolean
          default: false
        fuzzing_enabled:
          type: boolean
          default: false
        bug_bounty_mode:
          type: boolean
          default: false
        authorization_acknowledged:
          type: boolean
        authorization_note:
          type: string
        scanner_engines:
          type: array
          items:
            type: string
      required: [target_url, scan_profile, authorization_acknowledged]
    ScanJobListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ScanJob"
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
      required: [items, total, page, page_size]
    Finding:
      type: object
      properties:
        id:
          type: string
        scan_job_id:
          type: string
        title:
          type: string
        category:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        validation_status:
          type: string
          enum: [real, false_positive, informational, pending]
        confidence:
          type: number
        endpoint:
          type: string
        description:
          type: string
        evidence:
          type: string
        recommended_fix:
          type: string
        cvss_score:
          type: number
        cwe_id:
          type: string
        risk_score:
          type: number
        scanner_name:
          type: string
        created_at:
          type: string
        target_url:
          type: string
      required:
        - id
        - scan_job_id
        - title
        - category
        - severity
        - validation_status
        - confidence
        - endpoint
        - description
        - risk_score
        - scanner_name
        - created_at
...
    Target:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        domain:
          type: string
        last_scanned:
          type: string
        total_scans:
          type: integer
        total_findings:
          type: integer
        critical_findings:
          type: integer
        high_findings:
          type: integer
        risk_score:
          type: number
        status:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - url
        - domain
        - total_scans
        - total_findings
        - critical_findings
        - high_findings
        - risk_score
        - status
    TargetListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Target"
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
      required: [items, total, page, page_size]
    AttackSurfaceGraph:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
        metrics:
          $ref: "#/components/schemas/GraphMetrics"
      required: [nodes, edges]
    GraphNode:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [domain, endpoint, finding]
        severity:
          type: string
        metadata:
          type: object
      required: [id, label, type]
    GraphEdge:
      type: object
      properties:
        source:
          type: string
        target:
          type: string
        type:
          type: string
      required: [source, target, type]
    GraphMetrics:
      type: object
      properties:
        node_count:
          type: integer
        edge_count:
          type: integer
        finding_nodes:
          type: integer
        endpoint_nodes:
          type: integer
      required: [node_count, edge_count, finding_nodes, endpoint_nodes]
...